Rename the bullmq folder to workers.
